+++ config/sets/phpunit90.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\PHPUnit\CodeQuality\Rector\MethodCall\AssertRegExpRector;
+use Rector\PHPUnit\PHPUnit90\Rector\MethodCall\AssertRegExpRector;
 use Rector\PHPUnit\PHPUnit100\Rector\StmtsAwareInterface\WithConsecutiveRector;
 use Rector\PHPUnit\PHPUnit90\Rector\Class_\TestListenerToHooksRector;
 use Rector\PHPUnit\PHPUnit90\Rector\MethodCall\ExplicitPhpErrorApiRector;
