+++ rules-tests/PHPUnit90/Rector/MethodCall/AssertRegExpRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\PHPUnit\CodeQuality\Rector\MethodCall\AssertRegExpRector;
+use Rector\PHPUnit\PHPUnit90\Rector\MethodCall\AssertRegExpRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(AssertRegExpRector::class);
